Reload space after edit settings
